Create Check SLA OUT
This function is triggered after the user select two dates. Those two dates indicate the beginning and end of a time space defined by the user.
After that, will search in the database for incidents between these dates that broke the SLA(Service-Level Agreement), making a calculation based on the util time it took to restore the service.
All the incidents that broke the SLA are then stored in a table created for that purpose.